userspace/snmpd: make wrsPstatsHCTable to use 64 counters

+++ userspace/snmpd/wrsNetworkingStatusGroup.h
@@ -38,27 +38,27 @@ void init_wrsNetworkingStatusGroup(void);
 
 struct ns_pstats {
 	/* wrsEndpointStatus */
-	uint32_t TXUnderrun;		/* 1 */
-	uint32_t RXOverrun;		/* 2 */
-	uint32_t RXInvalidCode;		/* 3 */
-	uint32_t RXSyncLost;		/* 4 */
-	uint32_t RXPfilterDropped;	/* 6 */
-	uint32_t RXPCSErrors;		/* 7 */
-	uint32_t RXCRCErrors;		/* 10 */
+	uint64_t TXUnderrun;		/* 1 */
+	uint64_t RXOverrun;		/* 2 */
+	uint64_t RXInvalidCode;		/* 3 */
+	uint64_t RXSyncLost;		/* 4 */
+	uint64_t RXPfilterDropped;	/* 6 */
+	uint64_t RXPCSErrors;		/* 7 */
+	uint64_t RXCRCErrors;		/* 10 */
 	/* wrsSwcoreStatus */
 	/* Too much HP traffic / Per-priority queue full */
-	uint32_t RXFrames;		/* 20 */
-	uint32_t RXPrio0;		/* 22 */
-	uint32_t RXPrio1;		/* 23 */
-	uint32_t RXPrio2;		/* 24 */
-	uint32_t RXPrio3;		/* 25 */
-	uint32_t RXPrio4;		/* 26 */
-	uint32_t RXPrio5;		/* 27 */
-	uint32_t RXPrio6;		/* 28 */
-	uint32_t RXPrio7;		/* 29 */
-	uint32_t FastMatchPriority;	/* 33 */
+	uint64_t RXFrames;		/* 20 */
+	uint64_t RXPrio0;		/* 22 */
+	uint64_t RXPrio1;		/* 23 */
+	uint64_t RXPrio2;		/* 24 */
+	uint64_t RXPrio3;		/* 25 */
+	uint64_t RXPrio4;		/* 26 */
+	uint64_t RXPrio5;		/* 27 */
+	uint64_t RXPrio6;		/* 28 */
+	uint64_t RXPrio7;		/* 29 */
+	uint64_t FastMatchPriority;	/* 33 */
 	/* wrsRTUStatus */
-	uint32_t RXDropRTUFull;		/* 21 */
+	uint64_t RXDropRTUFull;		/* 21 */
 };
 
 /* parameters read from dot-config */
